feat: errorCodesからmessageにLOG_MESSAGESを移動し、詳細なエラーメッセージを定義
