+++ frontend/src/app/api/email/anon-post/route.ts
@@ -3,7 +3,7 @@ import { Resend } from 'resend'
 import { InternalError, Success } from '@/lib/api/response'
 import { createMailHTML } from '@/lib/api/utils'
 import { withLogger } from '@/lib/api/wrapper'
-import { LOG_MESSAGES } from '@/lib/api/errorCodes'
+import { LOG_MESSAGES } from '@/lib/log/message'
 
 const { RESEND_API_KEY, FROM_EMAIL, SUPPORT_EMAIL } = process.env
